+++ TextRPG/TextRPG/Monster/Monster.cs
@@ -1,17 +1,5 @@
 ﻿namespace TextRPG
 {
-    //캐릭터 인터페이스
-
-    public interface Character
-    {
-        int Level { get; set; }
-        string Name { get; set; }
-        int Health { get; set; }
-        int Attack { get; set; }
-        int Defence { get; set; }
-        bool IsDead { get; }
-        void TakeDamage(int damage);
-    }
